fix langfuse testing flow
